# backend/openapi.json
{
    "openapi": "3.1.1",
    "info": {
        "title": "Nook API",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:5999/"
        }
    ],
    "paths": {
...
        "/api/nodes/{id}/backlinks": {
            "get": {
                "tags": [
                    "Links"
                ],
                "description": "Links pointing at this page, including unresolved [[wikilinks]] that match its title or aliases.",
                "operationId": "GetBacklinks",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BacklinkDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/nodes/{id}/links": {
            "get": {
                "tags": [
                    "Links"
                ],
                "operationId": "GetOutgoingLinks",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OutgoingLinkDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/links/broken": {
            "get": {
                "tags": [
                    "Links"
                ],
                "operationId": "GetBrokenLinks",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "pattern": "^-?(?:0|[1-9]\\d*)$",
                            "type": [
                                "integer",
                                "string"
                            ],
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BrokenLinkDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/export": {
            "post": {
                "tags": [
                    "ExportImport"
                ],
                "description": "Streams a zip: <Title>.md next to a folder <Title>/ for children, frontmatter, relative internal links.",
                "operationId": "Export",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExportRequestDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/import": {
            "post": {
                "tags": [
                    "ExportImport"
                ],
                "description": "Imports .md/.markdown/.txt/.html/.csv/.zip. Archives over 20 MB return 202 { jobId }; poll GET /import/{jobId}.",
                "operationId": "Import",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "type": "object",
                                "allOf": [
                                    {
                                        "type": "object",
                                        "properties": {
                                            "file": {
                                                "$ref": "#/components/schemas/IFormFile"
                                            }
                                        }
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "parentId": {
                                                "type": "string",
                                                "format": "uuid"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImportResponse"
                                }
                            }
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImportAcceptedResponse"
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "Payload Too Large",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/import/{jobId}": {
            "get": {
                "tags": [
                    "ExportImport"
                ],
                "operationId": "GetImportJob",
                "parameters": [
                    {
                        "name": "jobId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImportJobStatusResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BacklinkDto": {
                "required": [
                    "sourceNode",
                    "blockId",
                    "kind",
                    "snippet"
                ],
                "type": "object",
                "properties": {
                    "sourceNode": {
                        "$ref": "#/components/schemas/NodeSummaryDto"
                    },
                    "blockId": {
                        "type": [
                            "null",
                            "string"
                        ],
                        "format": "uuid"
                    },
                    "kind": {
                        "type": "string"
                    },
                    "snippet": {
                        "type": "string"
                    }
                }
            },
...
            "BrokenLinkDto": {
                "required": [
                    "sourceNode",
                    "blockId",
                    "href",
                    "targetNodeId"
                ],
                "type": "object",
                "properties": {
                    "sourceNode": {
                        "$ref": "#/components/schemas/NodeSummaryDto"
                    },
                    "blockId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "href": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "targetNodeId": {
                        "type": [
                            "null",
                            "string"
                        ],
                        "format": "uuid"
                    }
                }
            },
...
            "ExportRequestDto": {
                "required": [
                    "nodeIds",
                    "includeChildren",
                    "format",
                    "includeFiles"
                ],
                "type": "object",
                "properties": {
                    "nodeIds": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    "includeChildren": {
                        "type": [
                            "null",
                            "boolean"
                        ]
                    },
                    "format": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "includeFiles": {
                        "type": [
                            "null",
                            "boolean"
                        ]
                    }
                }
            },
...
            "IFormFile": {
                "type": "string",
                "format": "binary"
            },
            "ImportAcceptedResponse": {
                "required": [
                    "jobId"
                ],
                "type": "object",
                "properties": {
                    "jobId": {
                        "type": "string",
                        "format": "uuid"
                    }
                }
            },
            "ImportJobStatusResponse": {
                "required": [
                    "jobId",
                    "status",
                    "result",
                    "error",
                    "createdAt",
                    "finishedAt"
                ],
                "type": "object",
                "properties": {
                    "jobId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "status": {
                        "type": "string"
                    },
                    "result": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/ImportResponse"
                            }
                        ]
                    },
                    "error": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "finishedAt": {
                        "type": [
                            "null",
                            "string"
                        ],
                        "format": "date-time"
                    }
                }
            },
            "ImportResponse": {
                "required": [
                    "pagesCreated",
                    "nodeIds",
                    "warnings"
                ],
                "type": "object",
                "properties": {
                    "pagesCreated": {
                        "pattern": "^-?(?:0|[1-9]\\d*)$",
                        "type": [
                            "integer",
                            "string"
                        ],
                        "format": "int32"
                    },
                    "nodeIds": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    "warnings": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "NodeIcon": {
                "required": [
                    "type",
                    "value"
                ],
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string"
                    },
                    "value": {
                        "type": "string"
                    }
                }
            },
...
            "NodeSummaryDto": {
                "required": [
                    "id",
                    "title",
                    "icon",
                    "kind",
                    "parentId"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "icon": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/NodeIcon"
                            }
                        ]
                    },
                    "kind": {
                        "type": "string"
                    },
                    "parentId": {
                        "type": [
                            "null",
                            "string"
                        ],
                        "format": "uuid"
                    }
                }
            },
...
            "OutgoingLinkDto": {
                "required": [
                    "targetNode",
                    "targetBlockId",
                    "kind",
                    "href",
                    "broken"
                ],
                "type": "object",
                "properties": {
                    "targetNode": {
                        "oneOf": [
                            {
                                "type": "null"
                            },
                            {
                                "$ref": "#/components/schemas/NodeSummaryDto"
                            }
                        ]
                    },
                    "targetBlockId": {
                        "type": [
                            "null",
                            "string"
                        ],
                        "format": "uuid"
                    },
                    "kind": {
                        "type": "string"
                    },
                    "href": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "broken": {
                        "type": "boolean"
                    }
                }
            },
...
            "ProblemDetails": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "title": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "status": {
                        "pattern": "^-?(?:0|[1-9]\\d*)$",
                        "type": [
                            "null",
                            "integer",
                            "string"
                        ],
                        "format": "int32"
                    },
                    "detail": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "instance": {
                        "type": [
                            "null",
                            "string"
                        ]
                    }
                }
            },
...
        }
    },
...
}